feat: added multiple stock functionality
- Stocks are now able to be trained based off of a basket of stocks and a target stock.

to use:
from stock_prediction import stock_prediction

stock_prediction():
-Parameters:

tickers: This is a list of stock symbols for which predictions will be made.
target_ticker: The main stock ticker symbol that we want to focus the prediction on.
num_future_minutes=4: Specifies how many minutes ahead we want to predict the stock price for. In this case, we're predicting for the next 4 minutes.
start_date_param=start: The start date from which we begin fetching stock data. This is provided as a start variable.
end_date_param=end: The end date for the stock data. This is provided as an end variable.

Functionality:

The function fetches historical stock data for the specified tickers within the given date range (start_date_param to end_date_param).
It then uses machine learning (likely an XGBoost model) to predict future stock prices based on the historical data.
The result is a JSON object containing the predicted stock prices for the next 4 minutes.